refactor: cursor와 page 디렉토리 경로 수정
