+++ dailyge-api/src/main/java/project/dailyge/app/core/retrospect/application/RetrospectReadService.java
@@ -1,7 +1,7 @@
 package project.dailyge.app.core.retrospect.application;
 
 import project.dailyge.app.core.common.auth.DailygeUser;
-import project.dailyge.app.page.CustomPageable;
+import project.dailyge.app.paging.CustomPageable;
 import project.dailyge.app.response.AsyncPagingResponse;
 import project.dailyge.entity.retrospect.RetrospectJpaEntity;
