+++ support/src/main/java/project/dailyge/app/paging/Cursor.java
@@ -1,4 +1,4 @@
-package project.dailyge.app.cursor;
+package project.dailyge.app.paging;
 
 import lombok.Getter;
