+++ support/src/main/java/project/dailyge/app/paging/CursorFactory.java
@@ -1,4 +1,4 @@
-package project.dailyge.app.cursor;
+package project.dailyge.app.paging;
 
 public final class CursorFactory {
