+++ support/src/main/java/project/dailyge/app/paging/PageFactory.java
@@ -1,4 +1,4 @@
-package project.dailyge.app.page;
+package project.dailyge.app.paging;
 
 public final class PageFactory {
